// Neptune.Web/src/app/shared/generated/model/funding-source-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundingSourceUpsertDto { 
    OrganizationID?: number;
    FundingSourceName?: string | null;
    IsActive?: boolean;
    FundingSourceDescription?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundingSourceUpsertDtoForm { 
    OrganizationID?: FormControl<number>;
    FundingSourceName?: FormControl<string>;
    IsActive?: FormControl<boolean>;
    FundingSourceDescription?: FormControl<string>;
}

export class FundingSourceUpsertDtoFormControls { 
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsActive = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}